extract MoviesRequestResponse into a separate class
